+++ db/mssql/QueryBuilder.php
@@ -17,6 +17,8 @@ use yii\base\InvalidParamException;
  */
 class QueryBuilder extends \yii\db\QueryBuilder
 {
+    protected $_oldMssql;
+
     /**
      * @var array mapping from abstract column types (keys) to physical column types (values).
      */
@@ -189,8 +191,7 @@ class QueryBuilder extends \yii\db\QueryBuilder
         $originalOrdering = $this->buildOrderBy($query->orderBy);
         if ($query->select) {
             $select = implode(', ', $query->select);
-        }
-        else {
+        } else {
             $select = $query->select = '*';
         }
         if ($select === '*') {
@@ -238,8 +239,11 @@ class QueryBuilder extends \yii\db\QueryBuilder
      */
     protected function isOldMssql()
     {
-        $pdo = $this->db->getSlavePdo();
-        $version = preg_split("/\./", $pdo->getAttribute(\PDO::ATTR_SERVER_VERSION));
-        return $version[0] < 11;
+        if ($this->_oldMssql === null) {
+            $pdo = $this->db->getSlavePdo();
+            $version = preg_split("/\./", $pdo->getAttribute(\PDO::ATTR_SERVER_VERSION));
+            $this->_oldMssql = $version[0] < 11;
+        }
+        return $this->_oldMssql;
     }
 }